feat: add meshery utils model to meshkit
