+++ utils/registry/spreadsheet_update.go
@@ -222,3 +222,12 @@ func marshalStructToCSValues[K any](data []*K) ([][]interface{}, error) {
 
 	return results, nil
 }
+
+func GetSheetIDFromTitle(s *sheets.Spreadsheet, title string) int64 {
+	for _, sheet := range s.Sheets {
+		if sheet.Properties.Title == title {
+			return sheet.Properties.SheetId
+		}
+	}
+	return -1
+}
